+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Added Coordination semaphore watch support for `DescribeSemaphore`: `options.WithWatchData`, `options.WithWatchOwners`, and `options.WithOnChanged` ([#2257](https://github.com/ydb-platform/ydb-go-sdk/issues/2257))
+## Unreleased
+* Added Coordination `DescribeSemaphore` one-shot watch via `options.WithSemaphoreWatch` ([#2257](https://github.com/ydb-platform/ydb-go-sdk/issues/2257))
 
 ## v3.146.0
 * Added `ydb.WithDefaultIdempotent(bool)` driver option